+++ packages/library/src/components.d.ts
@@ -200,9 +200,17 @@ export namespace Components {
          */
         "inverted": boolean;
         /**
+          * If `true` the navigation is handled by the component
+         */
+        "navigation": boolean;
+        /**
           * Select a step.
          */
         "select": (step: BalCardStepOption) => Promise<void>;
+        /**
+          * Hides the navigation circles and adds the step label instead
+         */
+        "showLabel": boolean;
     }
     interface BalCardSubtitle {
         /**
@@ -1155,6 +1163,10 @@ declare namespace LocalJSX {
          */
         "inverted"?: boolean;
         /**
+          * If `true` the navigation is handled by the component
+         */
+        "navigation"?: boolean;
+        /**
           * Emitted when the back button is clicked.
          */
         "onBalCardStepsBackClick"?: (event: CustomEvent<void>) => void;
@@ -1166,6 +1178,10 @@ declare namespace LocalJSX {
           * Emitted when the step circle is clicked.
          */
         "onBalCardStepsStepClick"?: (event: CustomEvent<BalCardStepOption>) => void;
+        /**
+          * Hides the navigation circles and adds the step label instead
+         */
+        "showLabel"?: boolean;
     }
     interface BalCardSubtitle {
         /**