+++ packages/library/src/components/bal-select-option/bal-select-option.type.ts
@@ -1,7 +1,7 @@
 export interface BalOptionValue<T> {
   value: T
   text: string
-  render?(data: any): string
+  render?(option: BalOptionValue<T>): string
 }
 
 export const NewBalOptionValue = <T>(value: T, text: string, render?: (data: any) => string) => {